添加自定义DrawerLayout，解决滑动冲突和之前的bug

+++ app/src/main/java/listview/example/x/slidelistview/SlideLayout.java
@@ -86,8 +86,8 @@ public class SlideLayout extends FrameLayout {
                 downY = startY = y;
                 break;
             case MotionEvent.ACTION_MOVE:
-                final float dx = (int) (x - startX);
-                final float dy = (int) (startY - y);
+                final float dx =  x - startX;
+                final float dy =  startY - y;
 
                 int disX = (int) (getScrollX() - dx);
                 if (disX <= 0) {
